Update Testing.py
so found the shape of the gray image and it will not be accepted into that function.
the cofuision began when i thought the process step in my notes regarding the cascade method as something that i had to do
but it was meant to be something that the function already does and is part of its integral procedure and not something that i have to do.
the function will have a 3 channel pic only.