Changing
so tried doing the xml conversion but that didn't work either, then just refered to the capstone notebook file to see hwat .xml file has he used but he used a different approach.
there i found out a way to isolate the hand from the background so will try that,
and he also used another approach for calculating the fingers, something that i might use alter in my projects of gesture controls and all but not in this capstone project.
i will use the idea that has come to me originally.