Update Testing.py
so this new approach does not include the dataset, this uses localized thresholding.
the concept is that therer is a specified area in which the segment of the frame will be thresholded to the hand will be detected and thresholded  here

then the baisc original idea will go on where the contout detection is done for the hand and the difference of frame is used to find the no. of contours.

furter testing is being done, then this will be evolved using the hand detecting dataset so that we can detet hand in any place inside a frame.