New Concept initiation
now test 3 file is made to test another theory
concept is that we might not create distancetranformation at 2 instances
as they create a lot of places for noise
i.e. that noise is stacked overtime and that lets the image be noisy, so just thresholded them at the 2 instances now the result has very less noise.

now it has to be seen how it does to distance transform now