Update Testing3.py
so this is the approach where i subtract gray scale of the background and the changes and then operate on them in the end. here noise was zero but the hand was not able to be detected at all, so wil now see how instructor has done this task of hand isolation
then i will isolate the hand and then will try my own logic to detect the counting on fingers.