Update Test5.py
so the logic is working, but the problem is with distance transformation function. it's not working at all and not even showing the pic.
but it's working for other functions and hence is able to isolate the fingers.

the problems here are
1) the distance transform function,
       it does not work with cont_img but works on thresholded image, idw why and it also doens't display the image it's supposed to.

2)there are black spots on the whole screen
i was able to isolate the image in result of test4 too but this is somewhat different.
3) contour detection for seprating the fingers in finger pic won't work becaueod these balck spots in the image. so will try to improve this.

so these are the 3 problems.
but the rest is working
now i can work on dissapeearing the balck spots and then i can work on finding the contours and operate on them using my logic for sepreation
black spots and contours are maiin issues the distance transform thingie is not so much troublesome as there is only problem in displaying.