Update Test5.py
this was the fnction with desired logic but the thing is ocont_img is still giving assertion error when triytin to use it with distance transformation.

but also found out that contour is not a good quality output as compared to out threshold
threshold is even better let alon the black spots, but still threshold is better so will use threshold only from now on

but as i didn't find any black spot removal solutions and reason for that is i didn't look for it so yeah, now i will learnt hte ways in video now, and i am dropping my idea, though it could have worked but only if the black spots would be removed, but still.