Update Test5.py
so did something that i learnt from the vid to solve the contour and distance tranformation problem

the problme had been the assertion error
meaning that the input array might be empty or has lesss dimensions than required by the function. so i used the bitwise_and operaton using the th as the base and the cont_img as it's mask. this will filter the thresholded image according to the contour mask that is the single contour and will be of same dimensions as that of the th, so will be able to work with distance transformation.
so this has become the final and most accurate to my original approach